changing commented out parts
