+++ models/user.js
@@ -1,8 +1,8 @@
 module.exports = function(sequelize, DataTypes) {
-    var User = sequelize.define("gitpet_users", {
+    var NewUser = sequelize.define("gitpet_users", {
         uname: DataTypes.STRING,
         email: DataTypes.STRING,
         pass: DataTypes.STRING
     });
-    return User;
+    return NewUser;
 };
