+++ public/index.html
@@ -163,6 +163,7 @@
         </div>
 
         <script src="/assets/js/index.js" type="text/javascript"></script>
+        <script src="/assets/js/newuser.js" type="text/javascript"></script>
 
 
     </body>
