+++ public/search.html
@@ -103,7 +103,7 @@
         </div>
 
         <div class="container" id="search-results">
-            
+
         </div>
 
         <div class="container">
@@ -123,7 +123,7 @@
         </div>
 
         <script src="/assets/js/index.js" type="text/javascript"></script>
-        <script src="/assets/js/search.js" type="text/javascript"></script>
+        <script src="/assets/js/newuser.js" type="text/javascript"></script>
 
 
     </body>
