+++ server.js
@@ -55,6 +55,7 @@
 
 var express = require("express");
 var exphbs = require("express-handlebars");
+var mysql2 = require("mysql2");
 
 var app = express();
 var PORT = process.env.PORT || 8080;
